requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
transformers
torch
opencv-python-headless
PyMuPDF
numpy
